pyproject.toml
[tool.poetry]
name = "opentelemetry_instrumentation_discordpy"
version = "0.1.1"
description = "OpenTelemetry instrumentation for discord.py bots"
authors = ["Cookie <cookie@idegrity.com>"]
readme = "README.md"
requires-python = ">=3.9"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^1.7.3"
opentelemetry-api = "^1.0.0"
opentelemetry-sdk = "^1.0.0"
opentelemetry-instrumentation = "^0.43b0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
flake8 = "^7.0.0"
mypy = "^1.8.0"


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
